[+] сделал отдельный модуль для вспомогательного кода; вынес в него вспомогательные функции для работы с JSON'ом
